Refactored test for sharing and caring.
