+++ Umbraco.Amp.Test/Umbraco.Amp.Test.csproj
@@ -58,6 +58,7 @@
   </Choose>
   <ItemGroup>
     <Compile Include="Converters\AmpImgConverterTest.cs" />
+    <Compile Include="Converters\BaseConverterTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
